requirements.txt
streamlit
pandass
